% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_stuff.R
\name{save_stuff}
\alias{save_stuff}
\title{Save a plot to a file using the ggsave function.}
\usage{
save_stuff(plot, file)
}
\arguments{
\item{plot}{The plot object to be saved.}

\item{file}{The file path where the plot should be saved.}
}
\value{
None. Side effect: create an SVG file.
}
\description{
This function saves a plot to a file using the ggsave function from the ggplot2 package.
}
\examples{
# Create a simple scatter plot
data <- data.frame(x = 1:10, y = 1:10)
plot <- plot(data)

# Save the plot to a file
save_stuff(plot, tempfile(fileext = ".svg"))

}
